[20260311] feat: add song and favorite management features

+++ infrastructure/postgres/init.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS songs (
     title VARCHAR(255) NOT NULL,
     year INT,
     language VARCHAR(10),
-    artist_id INT REFERENCES artists(id)
+    artist_id BIGINT REFERENCES artists(id)
 );
 
 -- Index pour la performance
